Add capabilties to create train/test split
fd_imputer can now open a dataset, split the dataset in test/train,
copy the dataset and test/train split into separate folders in a
folder called data/.

Also the train-set is copied to the folder where metanome reads its
datasets to detect FDs. This is cumbersome since FD detection has to be
started manually. An integration with metanome-cli would be superior
to this approach.