Add test for read_fds() and debug notebook
Code in fd_imputer_analysis.ipynb is buggy. This is due to the fact that
the data's index is treated differently by different components.

Now read_fds() treats the index as the 0th column, whilest split_df()
does not include the index as a column but as an index of a dataframe.

To solve this issue, better test-coverage is necessary. Also the
notebook's weight has to be reduced by moving the analysis into separate
functions.