Add random_dependency_generator, plot overfitting
Added a function to randomly generate dependencies. This is used to
test, how the overfitting classifier behaves on dependencies which are
not FDs compared to the standard classifier.

Added plots, comparing the overfitting classifier to the standard
classifier.

Re-ran the model-training and pickled the results. Results prior to this
commit might be faulty.