Extend Introduction, add Figures with Style&Color
Wrote some more text for the introduction. Also added the f1_ml_fd
comparison to the execution-part of the thesis.

Exported figures as .pdf to have better resultion. Improved color
settings. First tried fivethirtyeight, then chose seaborn-whitegrid.
Still not a final solution, still need to work on homogenous colors.

Remove colors package and replaced it with xcolors in main.tex. Still
need to come up with a way to use the same colors in the thesis as in
the graphs.