Add dependency detection and example classifier
A new file lib/dep_detector.py contains a class Dep_optimizer. It is
used for finding dependencies on a dataset. A top-down search pattern is
used, where a classifier for a full lhs is trained for each rhs
possible.

Then, only well-performing combinations are selected. In a next step,
every column is dropped once. Again, a classifier is trained for these
reduced lhs and again only the best combinations are selected.

This step is continued until minimal dependencies are found. Internally,
this is implemented on a tree (using the anytree lib).

The classifier_example is used for plotting the decision-boundary for
the theory part of the thesis.