Fix ml_imputer(), use mlfd's way to measure performance
Following Felix' suggestion, I chose the path of least resistance to
get the tests to pass. This now uses the old `Dataset` class as well as
`mlfd`'s way of measuring performance.

All test pass, although there are still some rather obvious type problems
with the contents of the tables that can be observed when `test_dep_detector`
is run.