Add splitting function for cleaning-experiments
Cleaning experiments require me to work with two relational instances
per dataset: a clean one and a dirty one.

In the PFD-approach, we train a model on the dirty data, and compare
y_pred to y_clean to get the cleaning performance, or y_dirty to get
the model performance.

The new split function thus saves train, test, validate, validate_clean
splits to the disk.

I might have to add dtype information to all datasets at some point --
just because this is crucial to how cleaning-experiments work. Remember
that "clean" data is the data that we define as "clean", and thus the
dtype is super curcial.