[wip] Fix cleaning experiment conceptually
I now carry out the cleaning experiment correctly. At the current state,
some debugging is still necessary. For one, I don't trust the result of
the error detection experiment. This should be working at least as well
as the cleaning experiment.

In addition, the predictor training appears to crash very frequently. I
need to investigate the origin of this bug and fix it.